pyproject.toml
[project]
name = "smart-qa-project"
version = "0.1.0"
description = "AI client for summarizing and querying texts."
authors = [
    {name = "Agwu Ekene"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-generativeai (>=0.8.5,<0.9.0)",
    "google-genai (>=1.53.0,<2.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "python-docx (>=1.2.0,<2.0.0)",
    "pypdf2 (>=3.0.1,<4.0.0)"
]

[dependency-groups]
dev = [
    "pytest (>=9.0.1,<10.0.0)",
    "pytest-mock (>=3.15.1,<4.0.0)",
    "pytest-cov (>=7.0.0,<8.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
